# openapi.yaml
openapi: 3.0.3
info:
  title: Unico Napp API
  version: 1.0.0
  description: |
    Documentação OpenAPI da API Unico Napp.
    Cada recurso contém explicações e exemplos de requisição/ resposta.
servers:
  - url: http://localhost:5000
    description: Servidor local de desenvolvimento
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    ErrorItem:
      type: object
      properties:
        codigo_erro:
          type: integer
          example: 1001
        descricao:
          type: string
          example: username - Possui valor inesperado ou vazio
    ErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorItem'
  responses:
...
    ValidationError:
      description: Payload validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            errors:
              - codigo_erro: 1001
                descricao: sku - Possui valor inesperado ou vazio
    ProductNotFound:
      description: Product was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            errors:
              - codigo_erro: 1013
                descricao: id_produto - produto não encontrado
...
    OrderNotFound:
      description: Order was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            errors:
              - codigo_erro: 1022
                descricao: Recurso não encontrado
...
paths:
...
  /api/v1/produtos:
    post:
      summary: Criar produto
      description: Cria um novo produto no catálogo.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreateRequest'
      responses:
        '201':
          description: Produto criado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  message:
                    type: string
                    example: created
        '400':
          $ref: '#/components/responses/ValidationError'
    get:
      summary: Listar produtos
      description: Retorna uma lista de produtos com suporte a paginação e filtros por SKU ou status.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: sku
          in: query
          schema:
            type: string
          description: Filtrar produto por SKU
        - name: status
          in: query
          schema:
            type: string
          description: Filtrar produto por status
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Offset da página
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: Limite de itens retornados
      responses:
        '200':
          description: Lista de produtos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
  /api/v1/produtos/{id}:
    get:
      summary: Consultar produto por ID
      description: Retorna os dados completos do produto, incluindo imagens e atributos.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Produto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/ProductNotFound'
    put:
      summary: Atualizar produto
      description: Substitui todos os campos do produto especificado.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreateRequest'
      responses:
        '200':
          description: Produto atualizado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: updated
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/ProductNotFound'
    patch:
      summary: Atualizar parcialmente produto
      description: Atualiza apenas os campos fornecidos no produto.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Produto atualizado parcialmente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: updated
        '404':
          $ref: '#/components/responses/ProductNotFound'
  /api/v1/orders:
    get:
      summary: Listar pedidos
      description: Retorna todos os pedidos com paginação.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Lista de pedidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
    post:
      summary: Criar pedido
      description: Cria um pedido associado a um produto e vendedor.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '201':
          description: Pedido criado
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  message:
                    type: string
                    example: created
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                validation_error:
                  value:
                    errors:
                      - codigo_erro: 1001
                        descricao: id_seller_contato - Possui valor inesperado ou vazio
                product_not_found:
                  value:
                    errors:
                      - codigo_erro: 1013
                        descricao: id_produto - produto não encontrado
  /api/v1/orders/{id}:
    get:
      summary: Consultar pedido por ID
      description: Retorna os detalhes de um pedido único.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Pedido encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          $ref: '#/components/responses/OrderNotFound'
    put:
      summary: Atualizar pedido
      description: Substitui todos os dados do pedido especificado.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '200':
          description: Pedido atualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: updated
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                validation_error:
                  value:
                    errors:
                      - codigo_erro: 1001
                        descricao: preco - Possui valor inesperado ou vazio
                product_not_found:
                  value:
                    errors:
                      - codigo_erro: 1013
                        descricao: id_produto - produto não encontrado
        '404':
          $ref: '#/components/responses/OrderNotFound'
    patch:
      summary: Atualizar parcialmente pedido
      description: Atualiza somente campos fornecidos no pedido.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Pedido atualizado parcialmente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: updated
        '400':
          $ref: '#/components/responses/ProductNotFound'
        '404':
          $ref: '#/components/responses/OrderNotFound'
  /api/v1/orders/seller/{seller_id}:
    get:
      summary: Pedidos por vendedor
      description: Retorna todos os pedidos associados a um vendedor.
      parameters:
        - name: seller_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Pedidos do vendedor
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
  /api/v1/orders/produto/{produto_id}:
    get:
      summary: Pedidos por produto
      description: Retorna pedidos associados a um produto.
      parameters:
        - name: produto_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Pedidos do produto
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
  /api/v1/orders/{id}/status:
    get:
      summary: Listar status de pedido
      description: Mostra todos os status registrados para um pedido.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de status do pedido
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderStatus'
    post:
      summary: Criar status de pedido
      description: Registra um novo status para um pedido e dispara webhook se configurado.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderStatusCreateRequest'
      responses:
        '200':
          description: Status criado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: created
                  id:
                    type: integer
                    example: 1
        '400':
          $ref: '#/components/responses/ValidationError'
    put:
      summary: Atualizar status de pedido
      description: Atualiza um status existente de pedido e dispara webhook de atualização.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderStatusCreateRequest'
      responses:
        '200':
          description: Status atualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: updated
        '404':
          $ref: '#/components/responses/OrderNotFound'
...